Bring back the shims
The .LNK workaround does not work well with Powershell, and it is a huge
hack anyway.

This time do things a bit differently. Five kinds of shims are provided,
with bundled logic to lookup installed Python, currently used Python,
SNAFU's bundled Python, and each of the first two has variants for
python and pip/easy_install (so we can provide hooking).

SNAFU's installation root and using versions are now stored in registry
so we can provide reliable lookup. That's a hit to portability, but I
don't think there is a better way around it.

A DLL for VC++ is now required in %PATH%, which saddens me. Alas, I
guess this is how Windows works.