Restructure the whole shim subproject
Now common logic are in a library so the Rust project settings are
cleaner. The code is significantly better and less duplicated as well.

The whole build process is wrapped as an Invoke task, so we don't need
to do so many pushd/popd or specifying --manifest-path.

snafu.exe is now built separately (from other shims), and are also
installed separately.

Some tests are added to verify the parsing process. Not comprehensive,
but make me much more confident with the code. The Docker tests are also
expanded a bit to cover more usages.