Update Rust dependencies
winreg -> 0.5.0
The default permission is now read-only, which allows the code to be
cleaned up a little.

winapi -> 0.3.2
The 0.3 upgrade breaks some API. I now need to explicit cast raw handler
to * winapi::ctypes::c_void, (which makes me a bit unconfortable), and
need to specify and use indivisual features now. Surprisingly this does
not reduce the built size in any slight bit. What the hell.